test: TestConfig 위치 변경 (#26)
