+++ oversweet-domain/src/main/java/com/depromeet/oversweet/domain/franchise/repository/FranchiseJpaRepository.java
@@ -7,5 +7,4 @@ import java.util.Optional;
 
 public interface FranchiseJpaRepository extends JpaRepository<FranchiseEntity, Long>, FranchiseEntityFilter {
 
-    Optional<FranchiseEntity> findById(Long franchiseId);
 }
