+++ oversweet-domain/src/test/resources/application.yml
@@ -1,0 +1,16 @@
+# JPA
+spring:
+  jpa:
+    properties:
+      hibernate:
+        format_sql: true # 포맷 형식
+        highlight_sql: true # sql 하이라이팅
+        default_batch_fetch_size: 1000
+        use_sql_comments: true # 개발자가 작성한 JPQL 확인 가능
+
+logging:
+  level:
+    org:
+      springframework.transaction.interceptor: trace # 트랜잭션 시작 여부 확인 가능
+      hibernate:
+        SQL: debug # SQL 쿼리 확인 가능 (JPQL X)
